backend/ppt/requirements.txt
crewai>=0.11.0
flask==3.1.0
python-dotenv==1.0.1
python-socketio==5.11.4
flask-socketio==5.4.1
google-generativeai==0.8.3
flask-cors==5.0.0
Pillow==10.4.0
requests==2.32.3
dataclasses-json==0.6.4
weasyprint==60.2
beautifulsoup4==4.12.3
